+++ lua/src/Lua/Primary.hs
@@ -27,6 +27,7 @@ to problems, then the package should be configured without flag
 -}
 module Lua.Primary
   ( lua_absindex
+  , lua_arith
   , lua_checkstack
   , lua_close
   , lua_concat
@@ -120,6 +121,45 @@ foreign import ccall unsafe "lua.h lua_absindex"
   lua_absindex :: Lua.State
                -> StackIndex     -- ^ idx
                -> IO StackIndex
+
+-- | Performs an arithmetic or bitwise operation over the two values (or
+-- one, in the case of negations) at the top of the stack, with the
+-- value at the top being the second operand, pops these values, and
+-- pushes the result of the operation. The function follows the
+-- semantics of the corresponding Lua operator (that is, it may call
+-- metamethods).
+--
+-- The value of @op@ must be one of the following constants:
+--
+-- -   __@LUA_OPADD@:__ performs addition (@+@)
+-- -   __@LUA_OPSUB@:__ performs subtraction (@-@)
+-- -   __@LUA_OPMUL@:__ performs multiplication (@*@)
+-- -   __@LUA_OPDIV@:__ performs float division (@\/@)
+-- -   __@LUA_OPIDIV@:__ performs floor division (@\/\/@)
+-- -   __@LUA_OPMOD@:__ performs modulo (@%@)
+-- -   __@LUA_OPPOW@:__ performs exponentiation (@^@)
+-- -   __@LUA_OPUNM@:__ performs mathematical negation (unary @-@)
+-- -   __@LUA_OPBNOT@:__ performs bitwise NOT (@~@)
+-- -   __@LUA_OPBAND@:__ performs bitwise AND (@&@)
+-- -   __@LUA_OPBOR@:__ performs bitwise OR (@|@)
+-- -   __@LUA_OPBXOR@:__ performs bitwise exclusive OR (@~@)
+-- -   __@LUA_OPSHL@:__ performs left shift (@\<\<@)
+-- -   __@LUA_OPSHR@:__ performs right shift (@>>@)
+--
+-- __WARNING__: @lua_arith@ is unsafe in Haskell: if the call to a
+-- metamethod triggers an error, then that error cannot be handled and
+-- will lead to an unrecoverable program crash. Consider using the
+-- @'Lua.hslua_arith'@ ersatz function instead. Likewise, the metamethod
+-- may not call a Haskell function unless the library was compiled
+-- without @allow-unsafe-gc@.
+--
+-- <https://www.lua.org/manual/5.3/manual.html#lua_arith>.
+foreign import ccall SAFTY "lua.h lua_arith"
+  lua_arith :: State -> ArithOPCode {- ^ op -} -> IO ()
+{-# WARNING lua_arith
+      [ "This is an unsafe function, errors will lead to a program crash;"
+      , "consider using hslua_arith instead."
+      ] #-}
 
 -- | Ensures that the stack has space for at least @n@ extra slots (that
 -- is, that you can safely push up to @n@ values into it). It returns